requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.5
gunicorn==20.1.0
idna==3.2
pkg_resources==0.0.0
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
responses==0.13.3
six==1.16.0
sqlparse==0.4.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.6
whitenoise==5.3.0
